\name{alpha}
\docType{data}
\alias{alpha}
\title{The alpha paramaters used to calculate PMA calls}
\description{
  The detection p value algorithm uses two parameters, alpha1 and alpha2 to define Present, MArginal Absent thresholds. These are chip specific and stored in this file.
}
\format{A CSV file, with two columns, Alpha1 and Alpha2 and a row for each chip.}
\keyword{datasets}
