\name{getQCParams-methods}
\docType{methods}
\alias{getQCParams-methods}
\alias{getQCParams,AffyBatch-method}
\alias{getQCParams}

\title{ Get Array Dependent QC Parameters }
\description{
  Gets Parameters associated with QC that are dependent on the particular array type being looked at. 
}
\section{Methods}{
\describe{

\item{x = "AffyBatch"}{ The AffyBatch object containing the arrays we care about }
}}

\keyword{methods}
