\name{qcs}
\docType{data}
\alias{qcs}
\title{an example QC Stats object}
\description{
  This datasets gives sample qc data for 25 array hgu133a
  comparison between two cell lines (MCF7 and MCF10A) and a variety of protocols.
}
\usage{qcs}
\format{a QCStats object}

\examples{
  data(qcs)
  plot(qcs)
}

\keyword{datasets}
