\name{spikeInProbes}
\alias{spikeInProbes}
\alias{qcProbes}
\alias{avbg}
\alias{maxbg}
\alias{minbg}
\alias{percent.present}
\alias{sfs}
\alias{target}
\alias{ratios}
\title{ Accessor Methods for QCStats class }
\description{
  Get and set my slots
}
\usage{
spikeInProbes(object)
qcProbes(object)
avbg(object)
maxbg(object)
minbg(object)
percent.present(object)
sfs(object)
target(object)
ratios(object)
}

\arguments{
  \item{object}{ the QCStats object you want to access }
}

\seealso{
  type \code{showMethods("spikeInProbes")} etc. to find out more about what these accessors do
}


\author{ Crispin J Miller }

\keyword{ misc }
