\name{setQCEnvironment}
\alias{setQCEnvironment}

\title{ Establish the appropriate QC environment for the specified array }
\description{
  Affymetrix define a series of QC parameters for their arrays. Many of
  these rely on specific probeset that differ between arrays and are
  used to calculate things like 3'/5' ratios. See \code{qc.affy}
  for more details. These functions are used to set up the appropriate
  QC environment for the specified array. This is done by loading a
  configuration file, either from the packages data directory, or from
  the specified path. See the package vignette for details of the config
  file's syntax.
}
\usage{
 setQCEnvironment(array,path=NULL)
}

\arguments{
  \item{array}{ This should be the 'clean' cdf name of the array as generated by
    \code{\link[affy]{cleancdfname}} in the affy package.} 
  \item{path}{ Path to the file. By default, checks the package's own
    data directory - only needed if a defininition file is being
    specified manually, as described in the vignette.} 
}

\details{
  The usual way to get the 'clean' cdfname is as follows:
  \code{cleancdfname(cdfName(eset))}, where \code{eset} is an
  \code{\link[affy:AffyBatch-class]{AffyBatch}} object.
}

\value{
  none.
}

\references{ http://bioinformatics.picr.man.ac.uk/ }
\author{ Crispin J Miller }

\seealso{
  \code{\link[simpleaffy]{qc}}
}
\examples{
  setQCEnvironment("hgu133plus2cdf")
  setQCEnvironment(cleancdfname("HG-U133_Plus_2"))

}
\keyword{ misc }
